+++ CHANGELOG.md
@@ -24,6 +24,7 @@
 
 - add ``once`` param to ``String#replace`` and ``Regex#replace`` methods
     - now only replace first occurrence of pattern
+- add ``start``, ``end`` method to ``RegexMatch`` type. now get offsets of match string
 
 #### LSP